fix(bpf): move Decode functions to decode.go so they compile without ebpf tag
